RAG/requirements.txt
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
langchain-openai>=0.1.0
openai>=1.30.1
chromadb>=0.4.24
tqdm
requests
python-dotenv